requirements.txt
pygame
python-chess
Pillow